NEW RELEASE: Update to R-4.0 and Bioconductor 3.11
This PR updates the images anvil-rstudio-base to base based on a new
image from the rocker group, i.e, rockerdev/tidyverse:4.0.0-ubuntu18.04.

The anvil-rstudio-bioconductor image inherits from anvil-rstudio-base
and includes R-4.0.0 and Bioconductor 3.11, along with some updated
packages.

The new set of docker images are:

1. us.gcr.io/anvil-gcr-public/anvil-rstudio-base:0.0.3

2. us.gcr.io/anvil-gcr-public/anvil-rstudio-bioconductor:0.0.5

Some changes made to files in README.md, VERSION in each of the
directories for the images to update the version number.

The CHANGELOG.md reflects the location and new "tag" of the docker
images in the google container registry

- Update TERRA_R_PLATFORM_BINARY_VERSION to `0.99.1`